+++ NEWS.md
@@ -1,3 +1,9 @@
+# supabaseR 1.0.1 (2026-04-22)
+
+## Documentation
+
+* Updated `README` with [CRAN](https://cran.r-project.org/web/packages/supabaseR/index.html) installation instructions as the primary installation method.
+
 # supabaseR 1.0.0 (2026-04-07)
 
 ## Breaking changes